+++ CHANGES.rst
@@ -28,6 +28,7 @@ Enhancements
 * ``Equal[]`` now compares complex against other numbers properly.
 * Improvements in handling products with infinite factors: ``0 Infinity``-> ``Indeterminate``, and ``expr Infinity``-> ``DirectedInfinite[expr]``
 * `StringSplit` not support lists as a first argument.
+* ``SetDelayed`` now accept several conditions impossed both at LHS as well as RHS.
 
 Bug fixes
 +++++++++